JG-28 Feat: Sidebar 내 컴포넌트 분리 및 네비게이션바 선택지에 따른 조건부 렌더링 구현
