+++ src/components/SideBar/index.jsx
@@ -1,74 +1,27 @@
-import BusStopItem from './components/BusStopItem';
-import CurrentLocationHeader from './components/CurrentLocationHeader';
+import { navigationBarState } from '@atoms/navigationBarState';
+import BusDetailInfo from '@components/BusDetailInfo';
+import BusStopsAround from '@components/BusStopsAround';
+import { CATEGORY } from '@constants/const';
+import { useRecoilState } from 'recoil';
+
 import FindBusInput from './components/FindBusInput';
 import * as S from './styles';
+
+const SIDE_BAR_MAP = {
+  [CATEGORY.HOME]: BusStopsAround,
+  [CATEGORY.FAVORITE]: () => <div>즐겨찾기</div>,
+  [CATEGORY.TIMETABLE]: () => <div>시간표</div>,
+  [CATEGORY.MYINFO]: () => <div>내 정보</div>,
+  [CATEGORY.BUSDETAILINFO]: BusDetailInfo,
+};
+
 function SideBar() {
-  const busStopList = [
-    {
-      busStopName: '서울역',
-      busStopNumber: '1234',
-      busDirection: '서울역 방향',
-    },
-    {
-      busStopName: '서울역',
-      busStopNumber: '1234',
-      busDirection: '서울역 방향',
-    },
-    {
-      busStopName: '서울역',
-      busStopNumber: '1234',
-      busDirection: '서울역 방향',
-    },
-    {
-      busStopName: '서울역',
-      busStopNumber: '1234',
-      busDirection: '서울역 방향',
-    },
-    {
-      busStopName: '서울역',
-      busStopNumber: '1234',
-      busDirection: '서울역 방향',
-    },
-    {
-      busStopName: '서울역',
-      busStopNumber: '1234',
-      busDirection: '서울역 방향',
-    },
-    {
-      busStopName: '서울역',
-      busStopNumber: '1234',
-      busDirection: '서울역 방향',
-    },
-    {
-      busStopName: '서울역',
-      busStopNumber: '1234',
-      busDirection: '서울역 방향',
-    },
-    {
-      busStopName: '서울역',
-      busStopNumber: '1234',
-      busDirection: '서울역 방향',
-    },
-    {
-      busStopName: '서울역',
-      busStopNumber: '1234',
-      busDirection: '서울역 방향',
-    },
-  ];
+  const [category] = useRecoilState(navigationBarState);
+
   return (
     <S.Wrapper>
       <FindBusInput />
-      <CurrentLocationHeader />
-      <S.BusStopItemWrapper>
-        {busStopList.map((busStop, index) => (
-          <BusStopItem
-            key={index}
-            busStopName={busStop.busStopName}
-            busStopNumber={busStop.busStopNumber}
-            busDirection={busStop.busDirection}
-          />
-        ))}
-      </S.BusStopItemWrapper>
+      {SIDE_BAR_MAP[category]()}
     </S.Wrapper>
   );
 }
